requirements.txt
# Web Server
fastapi
uvicorn[standard]
python-multipart
supabase==2.11.0

# --- Núcleo de IA y Agentes ---
crewai==1.7.2
openai>=1.83.0,<1.84.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# --- Servidor Web y Webhooks ---
Flask==3.0.2
gunicorn==21.2.0
requests==2.32.5

# --- Procesamiento de Datos y Google (Para el Dashboard) ---
google-api-python-client>=2.125.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
gspread>=6.1.0
pandas

# --- Utilidades ---
python-dotenv==1.1.1
PyYAML==6.0.1

# ORM y Migraciones
sqlalchemy>=2.0.0
alembic>=1.12.0
redis

# Driver para PostgreSQL (Psycopg 3 es el estándar actual)
psycopg[binary]
